Cleanup View code
